Clean up project card grid
