+++ src/components/pages/home/project-card-grid.js
@@ -1,0 +1,52 @@
+import React from 'react';
+import dlv from 'dlv';
+import { useStaticQuery, graphql } from 'gatsby';
+import styled from '@emotion/styled';
+import HoverCard from '../../dev-hub/hover-card';
+import Grid from '../../dev-hub/grid';
+import { size } from '../../dev-hub/theme';
+import { transformProjectStrapiData } from '../../../utils/transform-project-strapi-data';
+
+const GridContainer = styled('div')`
+    padding: ${size.medium};
+`;
+
+const homeFeaturedProjects = graphql`
+    query HomeFeaturedProjects {
+        allStrapiProjects(limit: 3) {
+            nodes {
+                ...ProjectFragment
+            }
+        }
+    }
+`;
+
+const ProjectCard = ({ project, ...props }) => (
+    <HoverCard to={project.slug} image={project.image_url} {...props}>
+        {project.name}
+    </HoverCard>
+);
+
+const ProjectCardGrid = () => {
+    const data = useStaticQuery(homeFeaturedProjects);
+    const projects = dlv(data, ['allStrapiProjects', 'nodes'], []);
+    const mappedProjects = projects.map(transformProjectStrapiData);
+    return (
+        <GridContainer>
+            <Grid
+                numCols={2}
+                layout={{
+                    rowSpan: [1],
+                    colSpan: [2, 1, 1],
+                }}
+                rowHeight="250px"
+            >
+                {mappedProjects.map(project => (
+                    <ProjectCard key={project.name} project={project} />
+                ))}
+            </Grid>
+        </GridContainer>
+    );
+};
+
+export default ProjectCardGrid;
